src/agent_framework/requirements.txt
# Microsoft Agent Framework - Core packages
agent-framework>=1.0.0b251001
agent-framework[azure-ai]>=1.0.0b251001

# Azure Identity for authentication
azure-identity>=1.17.0

# HTTP client for MCP
httpx>=0.27.0

# FastAPI for API server
fastapi>=0.110.0
uvicorn>=0.30.0

# Environment variables
python-dotenv>=1.0.0

# Optional: For advanced features
# azure-ai-projects>=1.0.0b10  # If you need direct AIProjectClient access
# azure-monitor-opentelemetry>=1.6.0
# azure-monitor-opentelemetry-exporter>=1.0.0b27
# opentelemetry-instrumentation-fastapi>=0.45b0
